test(slider): merge duplicate unit test file and add integration tests
