+++ slider/tests/test_slider_integration.py
@@ -1,0 +1,196 @@
+"""
+Integration tests for the Slider integration.
+
+No API key required — slider is pure local compute (python-pptx).
+These tests exercise real presentation generation through the action handlers.
+
+Run with:
+    pytest slider/tests/test_slider_integration.py -m integration
+"""
+
+import os
+import sys
+import importlib
+import importlib.util
+import base64
+
+_parent = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
+_deps = os.path.abspath(os.path.join(os.path.dirname(__file__), "../dependencies"))
+sys.path.insert(0, _parent)
+sys.path.insert(0, _deps)
+
+import pytest  # noqa: E402
+from unittest.mock import AsyncMock, MagicMock  # noqa: E402
+from autohive_integrations_sdk.integration import ResultType  # noqa: E402
+
+_original_cwd = os.getcwd()
+os.chdir(_parent)
+_spec = importlib.util.spec_from_file_location("slide_maker_mod_intg", os.path.join(_parent, "slide_maker.py"))
+_mod = importlib.util.module_from_spec(_spec)
+_spec.loader.exec_module(_mod)
+os.chdir(_original_cwd)
+sys.modules["slide_maker_mod_intg"] = _mod
+
+slide_maker = _mod.slide_maker
+
+pytestmark = pytest.mark.integration
+
+
+@pytest.fixture
+def live_context():
+    ctx = MagicMock(name="ExecutionContext")
+    ctx.fetch = AsyncMock(name="fetch")
+    ctx.auth = {}
+    return ctx
+
+
+class TestCreatePresentation:
+    async def test_creates_blank_presentation(self, live_context):
+        result = await slide_maker.execute_action("create_presentation", {}, live_context)
+
+        assert result.type == ResultType.ACTION
+        data = result.result.data
+        assert "presentation_id" in data
+        assert isinstance(data["presentation_id"], str)
+        assert data["slide_count"] >= 1
+        assert data["saved"] is True
+
+    async def test_returns_valid_pptx_binary(self, live_context):
+        result = await slide_maker.execute_action("create_presentation", {}, live_context)
+
+        data = result.result.data
+        file_obj = data["file"]
+        assert "content" in file_obj
+        assert file_obj["name"].endswith(".pptx")
+        assert file_obj["contentType"] == "application/vnd.openxmlformats-officedocument.presentationml.presentation"
+
+        # PPTX is a ZIP — verify magic bytes PK
+        raw = base64.b64decode(file_obj["content"] + "==")
+        assert raw[:2] == b"PK"
+
+    async def test_create_with_title_and_subtitle(self, live_context):
+        result = await slide_maker.execute_action(
+            "create_presentation",
+            {"title": "Integration Test Deck", "subtitle": "Generated by pytest"},
+            live_context,
+        )
+
+        assert result.type == ResultType.ACTION
+        assert result.result.data["saved"] is True
+        assert result.result.data["slide_count"] >= 1
+
+
+class TestAddElements:
+    async def test_add_text_element(self, live_context):
+        create_result = await slide_maker.execute_action("create_presentation", {}, live_context)
+        presentation_id = create_result.result.data["presentation_id"]
+        file_obj = create_result.result.data["file"]
+
+        result = await slide_maker.execute_action(
+            "add_elements",
+            {
+                "presentation_id": presentation_id,
+                "slide_index": 0,
+                "elements": [{"content": "Hello from integration tests"}],
+                "files": [file_obj],
+            },
+            live_context,
+        )
+
+        assert result.type == ResultType.ACTION
+        assert result.result.data["successfully_added"] >= 1
+        assert result.result.data["saved"] is True
+
+    async def test_add_bullet_list(self, live_context):
+        create_result = await slide_maker.execute_action("create_presentation", {}, live_context)
+        presentation_id = create_result.result.data["presentation_id"]
+        file_obj = create_result.result.data["file"]
+
+        bullets = "- Point one\n- Point two\n- Point three"
+        result = await slide_maker.execute_action(
+            "add_elements",
+            {
+                "presentation_id": presentation_id,
+                "slide_index": 0,
+                "elements": [{"content": bullets}],
+                "files": [file_obj],
+            },
+            live_context,
+        )
+
+        assert result.type == ResultType.ACTION
+        assert result.result.data["successfully_added"] >= 1
+
+
+class TestFindAndReplace:
+    async def test_replace_placeholder(self, live_context):
+        create_result = await slide_maker.execute_action(
+            "create_presentation",
+            {"title": "[Company Name]", "subtitle": "[Tagline]"},
+            live_context,
+        )
+        presentation_id = create_result.result.data["presentation_id"]
+        file_obj = create_result.result.data["file"]
+
+        result = await slide_maker.execute_action(
+            "find_and_replace",
+            {
+                "presentation_id": presentation_id,
+                "replacements": [
+                    {"find": "[Company Name]", "replace": "Autohive"},
+                    {"find": "[Tagline]", "replace": "Automate everything"},
+                ],
+                "files": [file_obj],
+            },
+            live_context,
+        )
+
+        assert result.type == ResultType.ACTION
+        assert "summary" in result.result.data
+        assert result.result.data["saved"] is True
+
+        # Verify output is still a valid PPTX
+        raw = base64.b64decode(result.result.data["file"]["content"] + "==")
+        assert raw[:2] == b"PK"
+
+
+class TestFullWorkflow:
+    async def test_create_add_content_verify_binary(self, live_context):
+        # Create
+        create_result = await slide_maker.execute_action(
+            "create_presentation",
+            {"title": "Q4 Report", "subtitle": "Integration Test"},
+            live_context,
+        )
+        assert create_result.type == ResultType.ACTION
+        presentation_id = create_result.result.data["presentation_id"]
+        file_obj = create_result.result.data["file"]
+
+        # Add a slide
+        add_slide_result = await slide_maker.execute_action(
+            "add_slide",
+            {"presentation_id": presentation_id, "files": [file_obj]},
+            live_context,
+        )
+        assert add_slide_result.type == ResultType.ACTION
+        assert add_slide_result.result.data["slide_count"] == 2
+        file_obj = add_slide_result.result.data["file"]
+
+        # Add elements to slide 1
+        add_result = await slide_maker.execute_action(
+            "add_elements",
+            {
+                "presentation_id": presentation_id,
+                "slide_index": 1,
+                "elements": [{"content": "**Key Metrics**"}, {"content": "- Revenue: $1M\n- Users: 10k"}],
+                "files": [file_obj],
+            },
+            live_context,
+        )
+        assert add_result.type == ResultType.ACTION
+        file_obj = add_result.result.data["file"]
+
+        # Verify final binary is a valid PPTX
+        raw = base64.b64decode(file_obj["content"] + "==")
+        assert raw[:2] == b"PK"
+        assert file_obj["name"].endswith(".pptx")
